Free unused cal memory memory
This makes all the string in events and todos dynamically allocated.
The memory for the event and any underlying data is now owned by cal.c,
and will be freed when no longer needed.

Maybe someday we'll go back to making these constants, since that would
require less duplicated strings, but for now it's easiest to just make
everything freeable