Rework how the current window is stored
The current window is now a pointer to the actual window instead of an
index into the menu. The menu is also stored separate from full list of
windows and each window is created as a separate structure so that it
can be referenced by name.

The down side is that we can't use an 'enum' for config parsing, but on
the other hand, we're not duplicating the list of names at least.